Optimization tests 1276 (#1294)
* made a few small bug fixes

* removed a join

* removed bad unique name code #1276

* updated the generateOWnerString code for #1276

* minor fix

* minimized some of the getFeatureLocation calls

* removed un-necessary deseralization of id

+++ grails-app/services/org/bbop/apollo/FeatureService.groovy
@@ -183,7 +183,7 @@ class FeatureService {
                         tmpTranscript.name = nameService.generateUniqueName(tmpTranscript, tmpGene.name)
                     }
 
-                    if (overlapperService.overlaps(tmpTranscript, tmpGene)) {
+                    if (tmpTranscript && tmpGene && overlapperService.overlaps(tmpTranscript, tmpGene)) {
                         log.debug "There is an overlap, adding to an existing gene"
                         transcript = tmpTranscript;
                         gene = tmpGene;
@@ -1582,18 +1582,17 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
     }
 
     String generateOwnerString(Feature feature){
-        String finalOwnerString
+        if(feature.owner){
+          return feature.owner.username
+        }
         if (feature.owners) {
             String ownerString = ""
             for (owner in feature.owners) {
-                ownerString += feature.owner.username + " "
+                ownerString += owner.username + " "
             }
-            finalOwnerString = ownerString?.trim()
-        } else if (feature.owner) {
-            finalOwnerString = feature?.owner?.username
-        } else {
-            finalOwnerString = "None"
+            return ownerString
         }
+        return "None"
     }
 
     /**
@@ -1988,7 +1987,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
         ArrayList<Transcript> transcriptsToUpdate = new ArrayList<Transcript>()
 
         for (Transcript eachTranscript : allSortedTranscripts) {
-            if (overlapperService.overlaps(eachTranscript, fivePrimeGene)) {
+            if (eachTranscript && fivePrimeGene && overlapperService.overlaps(eachTranscript, fivePrimeGene)) {
                 if (transcriptService.getGene(eachTranscript).uniqueName != fivePrimeGene.uniqueName) {
                     transcriptsToAssociate.add(eachTranscript)
                     genesToMerge.add(transcriptService.getGene(eachTranscript))
@@ -2058,7 +2057,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
                     firstTranscript.save(flush: true)
                     continue
                 }
-                if (overlapperService.overlaps(eachTranscript, firstTranscript)) {
+                if (eachTranscript && firstTranscript && overlapperService.overlaps(eachTranscript, firstTranscript)) {
                     featureRelationshipService.removeFeatureRelationship(transcriptService.getGene(eachTranscript), eachTranscript)
                     addTranscriptToGene(transcriptService.getGene(firstTranscript), eachTranscript)
                     firstTranscript.name = nameService.generateUniqueName(firstTranscript, transcriptService.getGene(firstTranscript).name)
@@ -2076,7 +2075,7 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
         overlappingTranscripts.remove(transcript) // removing itself
         ArrayList<Transcript> transcriptsWithOverlappingOrf = new ArrayList<Transcript>()
         for (Transcript eachTranscript in overlappingTranscripts) {
-            if (overlapperService.overlaps(eachTranscript, transcript)) {
+            if (eachTranscript && transcript && overlapperService.overlaps(eachTranscript, transcript)) {
                 transcriptsWithOverlappingOrf.add(eachTranscript)
             }
         }
